requirements.txt
pycrypto==2.6.1
gmpy2==2.0.3
numpy==1.9.0
sphinx==1.2.3
sphinxcontrib-napoleon==0.2.8
nose==1.3.4
